+++ webtrees/CHANGELOG.md
@@ -1,3 +1,9 @@
+- Allow Mariadb addon autodiscovery as database
+
+## 2.1.0-alpha.1 (28-12-2021)
+- Switch to 2.1
+- Update to latest version from nathanvaughn/webtrees-docker
+- Update to latest version from nathanvaughn/webtrees-docker
 - New standardized logic for Dockerfile build and packages installation
 
 ## 2.0.19 (07-12-2021)